# shared/schemas/api.openapi.json
{
  "components": {
    "schemas": {
...
      "AppliesTo": {
        "properties": {
          "abend_code": {
            "title": "Abend Code",
            "type": "string"
          },
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph"
          },
          "program": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Program"
          }
        },
        "required": [
          "abend_code"
        ],
        "title": "AppliesTo",
        "type": "object"
      },
...
      "Envelope_RunbookApplyResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunbookApplyResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RunbookApplyResponse]",
        "type": "object"
      },
      "Envelope_RunbookCreateResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunbookCreateResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RunbookCreateResponse]",
        "type": "object"
      },
      "Envelope_RunbookListResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunbookListResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[RunbookListResponse]",
        "type": "object"
      },
...
      "Envelope_Runbook_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Runbook"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[Runbook]",
        "type": "object"
      },
...
      "Envelope_ScoreOverrideResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScoreOverrideResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[ScoreOverrideResponse]",
        "type": "object"
      },
      "Envelope_ScoreResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ScoreResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[ScoreResponse]",
        "type": "object"
      },
...
      "Envelope_WeightsResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WeightsResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[WeightsResponse]",
        "type": "object"
      },
      "Envelope_WeightsUpdateResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WeightsUpdateResponse"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[WeightsUpdateResponse]",
        "type": "object"
      },
...
      "FixAction": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "type": {
            "$ref": "#/components/schemas/FixActionType"
          }
        },
        "required": [
          "label",
          "type"
        ],
        "title": "FixAction",
        "type": "object"
      },
      "FixActionType": {
        "enum": [
          "sql",
          "jcl",
          "shell",
          "manual"
        ],
        "title": "FixActionType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunbookApplyRequest": {
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "RunbookApplyRequest",
        "type": "object"
      },
      "RunbookApplyResponse": {
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "runbook_id": {
            "title": "Runbook Id",
            "type": "string"
          }
        },
        "required": [
          "runbook_id",
          "application_id"
        ],
        "title": "RunbookApplyResponse",
        "type": "object"
      },
      "RunbookCreateRequest": {
        "properties": {
          "applies_to": {
            "items": {
              "$ref": "#/components/schemas/AppliesTo"
            },
            "minItems": 1,
            "title": "Applies To",
            "type": "array"
          },
          "body_markdown": {
            "minLength": 10,
            "title": "Body Markdown",
            "type": "string"
          },
          "fix_actions": {
            "items": {
              "$ref": "#/components/schemas/FixAction"
            },
            "title": "Fix Actions",
            "type": "array"
          },
          "title": {
            "maxLength": 200,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "applies_to",
          "body_markdown"
        ],
        "title": "RunbookCreateRequest",
        "type": "object"
      },
      "RunbookCreateResponse": {
        "properties": {
          "audit_event_id": {
            "title": "Audit Event Id",
            "type": "string"
          },
          "runbook": {
            "$ref": "#/components/schemas/Runbook"
          }
        },
        "required": [
          "runbook",
          "audit_event_id"
        ],
        "title": "RunbookCreateResponse",
        "type": "object"
      },
      "RunbookListResponse": {
        "properties": {
          "runbooks": {
            "items": {
              "$ref": "#/components/schemas/Runbook"
            },
            "title": "Runbooks",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "runbooks",
          "total"
        ],
        "title": "RunbookListResponse",
        "type": "object"
      },
...
      "ScoreOverrideRequest": {
        "properties": {
          "new_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "New Score",
            "type": "integer"
          },
          "reason": {
            "maxLength": 400,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "new_score",
          "reason"
        ],
        "title": "ScoreOverrideRequest",
        "type": "object"
      },
...
      "ScoreRequest": {
        "properties": {
          "jcl_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jcl Name"
          },
          "jcl_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jcl Source"
          },
          "region": {
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "region"
        ],
        "title": "ScoreRequest",
        "type": "object"
      },
      "ScoreResponse": {
        "properties": {
          "breakdown": {
            "$ref": "#/components/schemas/ScoreBreakdown"
          },
          "score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "score",
          "breakdown"
        ],
        "title": "ScoreResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WeightsUpdateRequest": {
        "properties": {
          "reason": {
            "maxLength": 400,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          },
          "weights": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Weights",
            "type": "object"
          }
        },
        "required": [
          "weights",
          "reason"
        ],
        "title": "WeightsUpdateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI control plane for z/OS modernization. See docs/API.md.",
    "title": "Helios Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runbooks": {
      "get": {
        "operationId": "list_runbooks_api_runbooks_get",
        "parameters": [
          {
            "in": "query",
            "name": "abend_code",
            "required": true,
            "schema": {
              "title": "Abend Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "program",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Program"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunbookListResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runbooks",
        "tags": [
          "runbooks"
        ]
      },
      "post": {
        "operationId": "create_runbook_api_runbooks_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunbookCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunbookCreateResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Runbook",
        "tags": [
          "runbooks"
        ]
      }
    },
    "/api/runbooks/{runbook_id}": {
      "get": {
        "operationId": "get_runbook_api_runbooks__runbook_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "runbook_id",
            "required": true,
            "schema": {
              "title": "Runbook Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Runbook_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runbook",
        "tags": [
          "runbooks"
        ]
      }
    },
    "/api/runbooks/{runbook_id}/apply": {
      "post": {
        "operationId": "apply_runbook_api_runbooks__runbook_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "runbook_id",
            "required": true,
            "schema": {
              "title": "Runbook Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunbookApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_RunbookApplyResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Runbook",
        "tags": [
          "runbooks"
        ]
      }
    },
...
    "/api/score": {
      "post": {
        "operationId": "compute_score_api_score_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ScoreResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Score",
        "tags": [
          "score"
        ]
      }
    },
    "/api/score/weights/{region}": {
      "get": {
        "operationId": "get_weights_api_score_weights__region__get",
        "parameters": [
          {
            "in": "path",
            "name": "region",
            "required": true,
            "schema": {
              "title": "Region",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_WeightsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Weights",
        "tags": [
          "score"
        ]
      },
      "post": {
        "operationId": "update_weights_api_score_weights__region__post",
        "parameters": [
          {
            "in": "path",
            "name": "region",
            "required": true,
            "schema": {
              "title": "Region",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeightsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_WeightsUpdateResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Weights",
        "tags": [
          "score"
        ]
      }
    },
    "/api/score/{event_id}/override": {
      "post": {
        "operationId": "override_score_api_score__event_id__override_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ScoreOverrideResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Override Score",
        "tags": [
          "score"
        ]
      }
    },
...
  }
}